feat: add configurable send shortcut
