+++ src-tauri/src/app_settings.rs
@@ -11,8 +11,20 @@ use crate::storage::atomic_write;
 use parking_lot::Mutex;
 use serde::{Deserialize, Serialize};
 
+fn default_send_shortcut() -> String {
+    "mod_enter".to_string()
+}
+
+fn normalize_send_shortcut(value: String) -> String {
+    match value.as_str() {
+        "enter" | "mod_enter" => value,
+        _ => default_send_shortcut(),
+    }
+}
+
 static CACHED_CLAUDE_VERSION: OnceLock<Mutex<Option<Option<String>>>> = OnceLock::new();
 static CACHED_CODEX_VERSION: OnceLock<Mutex<Option<Option<String>>>> = OnceLock::new();
+static SETTINGS_LOCK: OnceLock<Mutex<()>> = OnceLock::new();
 
 pub fn get_login_shell_env() -> &'static [(String, String)] {
     crate::platform::login_shell_env()
@@ -22,12 +34,24 @@ pub fn get_login_shell_path() -> &'static str {
     crate::platform::login_shell_path()
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, Default, PartialEq, Eq)]
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
 pub struct AppSettings {
     #[serde(default)]
     pub claude_path: String,
     #[serde(default)]
     pub codex_path: String,
+    #[serde(default = "default_send_shortcut")]
+    pub send_shortcut: String,
+}
+
+impl Default for AppSettings {
+    fn default() -> Self {
+        Self {
+            claude_path: String::new(),
+            codex_path: String::new(),
+            send_shortcut: default_send_shortcut(),
+        }
+    }
 }
 
 #[derive(Clone, Debug, Default)]
@@ -62,6 +86,10 @@ fn clear_cached_versions() {
     *CACHED_CODEX_VERSION
         .get_or_init(|| Mutex::new(None))
         .lock() = None;
+}
+
+fn settings_lock() -> &'static Mutex<()> {
+    SETTINGS_LOCK.get_or_init(|| Mutex::new(()))
 }
 
 fn nezha_dir() -> Result<PathBuf, String> {
@@ -277,10 +305,11 @@ fn normalize_settings(settings: AppSettings) -> AppSettings {
     AppSettings {
         claude_path: resolve_agent_launch_spec_from_path("claude", &settings.claude_path).program,
         codex_path: resolve_agent_launch_spec_from_path("codex", &settings.codex_path).program,
+        send_shortcut: normalize_send_shortcut(settings.send_shortcut),
     }
 }
 
-pub fn load_settings_internal() -> AppSettings {
+fn load_settings_unlocked() -> AppSettings {
     let path = match settings_path() {
         Ok(p) => p,
         Err(_) => return AppSettings::default(),
@@ -290,6 +319,7 @@ pub fn load_settings_internal() -> AppSettings {
         let settings = normalize_settings(AppSettings {
             claude_path: detect_path("claude"),
             codex_path: detect_path("codex"),
+            send_shortcut: default_send_shortcut(),
         });
         if let Ok(dir) = nezha_dir() {
             let _ = fs::create_dir_all(&dir);
@@ -314,33 +344,88 @@ pub fn load_settings_internal() -> AppSettings {
     normalized
 }
 
+pub fn load_settings_internal() -> AppSettings {
+    let _guard = settings_lock().lock();
+    load_settings_unlocked()
+}
+
 pub fn get_agent_launch_spec(agent: &str) -> AgentLaunchSpec {
     get_agent_launch_spec_from_settings(&load_settings_internal(), agent)
 }
 
 #[tauri::command]
-pub fn load_app_settings() -> Result<AppSettings, String> {
-    Ok(load_settings_internal())
+pub async fn load_app_settings() -> Result<AppSettings, String> {
+    tokio::task::spawn_blocking(load_settings_internal)
+        .await
+        .map_err(|e| e.to_string())
 }
 
 #[tauri::command]
 pub fn save_app_settings(settings: AppSettings) -> Result<(), String> {
-    let dir = nezha_dir()?;
-    fs::create_dir_all(&dir).map_err(|e| e.to_string())?;
-    let path = settings_path()?;
-    let normalized = normalize_settings(settings);
-    let raw = serde_json::to_string_pretty(&normalized).map_err(|e| e.to_string())?;
-    atomic_write(&path, &raw)?;
+    {
+        let _guard = settings_lock().lock();
+        let dir = nezha_dir()?;
+        fs::create_dir_all(&dir).map_err(|e| e.to_string())?;
+        let path = settings_path()?;
+        let normalized = normalize_settings(settings);
+        let raw = serde_json::to_string_pretty(&normalized).map_err(|e| e.to_string())?;
+        atomic_write(&path, &raw)?;
+    }
     clear_cached_versions();
     Ok(())
 }
 
 #[tauri::command]
-pub fn detect_agent_paths() -> Result<AppSettings, String> {
-    Ok(normalize_settings(AppSettings {
-        claude_path: detect_path("claude"),
-        codex_path: detect_path("codex"),
-    }))
+pub async fn save_agent_paths(claude_path: String, codex_path: String) -> Result<AppSettings, String> {
+    let normalized = tokio::task::spawn_blocking(move || {
+        let _guard = settings_lock().lock();
+        let mut settings = load_settings_unlocked();
+        settings.claude_path = claude_path;
+        settings.codex_path = codex_path;
+
+        let dir = nezha_dir()?;
+        fs::create_dir_all(&dir).map_err(|e| e.to_string())?;
+        let path = settings_path()?;
+        let normalized = normalize_settings(settings);
+        let raw = serde_json::to_string_pretty(&normalized).map_err(|e| e.to_string())?;
+        atomic_write(&path, &raw)?;
+        Ok::<AppSettings, String>(normalized)
+    })
+    .await
+    .map_err(|e| e.to_string())??;
+    clear_cached_versions();
+    Ok(normalized)
+}
+
+#[tauri::command]
+pub async fn save_send_shortcut(send_shortcut: String) -> Result<AppSettings, String> {
+    tokio::task::spawn_blocking(move || {
+        let _guard = settings_lock().lock();
+        let mut settings = load_settings_unlocked();
+        settings.send_shortcut = normalize_send_shortcut(send_shortcut);
+
+        let dir = nezha_dir()?;
+        fs::create_dir_all(&dir).map_err(|e| e.to_string())?;
+        let path = settings_path()?;
+        let normalized = normalize_settings(settings);
+        let raw = serde_json::to_string_pretty(&normalized).map_err(|e| e.to_string())?;
+        atomic_write(&path, &raw)?;
+        Ok::<AppSettings, String>(normalized)
+    })
+    .await
+    .map_err(|e| e.to_string())?
+}
+
+#[tauri::command]
+pub async fn detect_agent_paths() -> Result<AppSettings, String> {
+    tokio::task::spawn_blocking(|| {
+        let mut settings = load_settings_internal();
+        settings.claude_path = detect_path("claude");
+        settings.codex_path = detect_path("codex");
+        Ok(normalize_settings(settings))
+    })
+    .await
+    .map_err(|e| e.to_string())?
 }
 
 fn detect_version(launch: &AgentLaunchSpec) -> Option<String> {
@@ -420,16 +505,20 @@ pub fn claude_version_gte(saved_version: &str, min_version: &str) -> bool {
 }
 
 #[tauri::command]
-pub fn detect_agent_versions() -> Result<AgentVersions, String> {
-    Ok(AgentVersions {
+pub async fn detect_agent_versions() -> Result<AgentVersions, String> {
+    tokio::task::spawn_blocking(|| AgentVersions {
         claude_version: detect_claude_version().unwrap_or_default(),
         codex_version: detect_codex_version().unwrap_or_default(),
     })
+    .await
+    .map_err(|e| e.to_string())
 }
 
 #[tauri::command]
-pub fn detect_agent_versions_for_settings(settings: AppSettings) -> Result<AgentVersions, String> {
-    Ok(detect_versions_for_settings(&settings))
+pub async fn detect_agent_versions_for_settings(settings: AppSettings) -> Result<AgentVersions, String> {
+    tokio::task::spawn_blocking(move || detect_versions_for_settings(&settings))
+        .await
+        .map_err(|e| e.to_string())
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, Default)]
