+++ src/components/NewTaskView.tsx
@@ -17,6 +17,13 @@ import {
 import { ImageAttachments } from "./new-task/ImageAttachments";
 import { AgentPermSelector } from "./new-task/AgentPermSelector";
 import { useI18n } from "../i18n";
+import { APP_PLATFORM } from "../platform";
+import {
+  DEFAULT_SEND_SHORTCUT,
+  getSendShortcutKeys,
+  normalizeSendShortcut,
+  type SendShortcut,
+} from "../shortcuts";
 import claudeGif from "../assets/gif/claude.gif";
 import codexGif from "../assets/gif/codex.gif";
 import s from "../styles";
@@ -92,6 +99,7 @@ export function NewTaskView({
       !(initialDraft?.promptHtml ?? "").replace(/<[^>]+>/g, "").trim() &&
       (initialDraft?.pastedImages.length ?? 0) === 0,
   );
+  const [sendShortcut, setSendShortcut] = useState<SendShortcut>(DEFAULT_SEND_SHORTCUT);
 
   const { editorRef, isComposingRef, handle: editorHandle } = usePromptEditor();
   const editorContentRef = useRef<PromptEditorContent>({
@@ -142,6 +150,18 @@ export function NewTaskView({
       });
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  useEffect(() => {
+    function loadSendShortcut() {
+      invoke<{ send_shortcut?: string }>("load_app_settings")
+        .then((settings) => setSendShortcut(normalizeSendShortcut(settings.send_shortcut)))
+        .catch(() => setSendShortcut(DEFAULT_SEND_SHORTCUT));
+    }
+
+    loadSendShortcut();
+    window.addEventListener("nezha:app-settings-changed", loadSendShortcut);
+    return () => window.removeEventListener("nezha:app-settings-changed", loadSendShortcut);
   }, []);
 
   // Load default agent and permission mode from project config when project changes
@@ -401,6 +421,7 @@ export function NewTaskView({
             setMentionIndex(0);
           }}
           onSetMentionIndex={setMentionIndex}
+          sendShortcut={sendShortcut}
           onSubmit={handleSubmit}
           onContentChange={(content) => {
             editorContentRef.current = content;
@@ -420,6 +441,7 @@ export function NewTaskView({
           planMode={planMode}
           isEmpty={isEmpty}
           hasImages={pastedImages.length > 0}
+          sendShortcutKeys={getSendShortcutKeys(sendShortcut, APP_PLATFORM)}
           onSetAgent={setAgent}
           onSetPermMode={setPermMode}
           onTogglePlanMode={() => setPlanMode((v) => !v)}
