+++ src/i18n.tsx
@@ -17,6 +17,7 @@ const translations: Record<AppLanguage, Record<string, string>> = {
     "appSettings.title": "App Settings",
     "appSettings.general": "General",
     "appSettings.theme": "Theme",
+    "appSettings.shortcuts": "Shortcuts",
     "appSettings.about": "About",
     "appSettings.language": "Language",
     "appSettings.appLanguage": "App Language",
@@ -31,13 +32,21 @@ const translations: Record<AppLanguage, Record<string, string>> = {
     "appSettings.claudePathHint": "Leave empty to use claude from the system PATH.",
     "appSettings.codexPathHint": "Leave empty to use codex from the system PATH.",
     "appSettings.installedVersions": "Installed Versions",
-    "appSettings.versionsHint": "Versions are detected from the configured executable path or the system PATH.",
+    "appSettings.versionsHint":
+      "Versions are detected from the configured executable path or the system PATH.",
     "appSettings.nezhaDescription": "Desktop task manager for AI coding agents",
     "appSettings.version": "Version",
     "appSettings.github": "GitHub",
-    "appSettings.starHint": "If you think NeZha is helpful, please consider starring this project on GitHub.",
-    "appSettings.syntaxHighlightUnavailable": "Syntax highlighting is unavailable. Showing plain text instead.",
+    "appSettings.starHint":
+      "If you think NeZha is helpful, please consider starring this project on GitHub.",
+    "appSettings.syntaxHighlightUnavailable":
+      "Syntax highlighting is unavailable. Showing plain text instead.",
     "appSettings.configFileNotFound": "Config file not found: {path}",
+    "appSettings.sendMessage": "Send message",
+    "appSettings.sendShortcutModEnter": "Cmd/Ctrl + Enter",
+    "appSettings.sendShortcutModEnterHint": "Send with {shortcut}. Enter inserts a new line.",
+    "appSettings.sendShortcutEnter": "Enter",
+    "appSettings.sendShortcutEnterHint": "Send with Enter. Cmd/Ctrl + Enter inserts a new line.",
     "theme.followSystem": "Follow System",
     "theme.followSystemAria": "Follow system theme",
     "theme.manualTheme": "Manual Theme",
@@ -135,26 +144,28 @@ const translations: Record<AppLanguage, Record<string, string>> = {
     "task.failed": "Task failed",
     "task.cancelled": "Task cancelled",
     "task.deleteTitle": "Delete Task",
-    "task.deletePrompt": "Delete task \"{prompt}\"?",
+    "task.deletePrompt": 'Delete task "{prompt}"?',
     "task.clearTitle": "Clear Tasks",
     "task.clearPrompt": "Delete all {count} tasks in {project}?",
     "task.deleteProjectTitle": "Delete Project",
-    "task.deleteProjectPrompt": "Delete project \"{project}\" and all its task records?",
+    "task.deleteProjectPrompt": 'Delete project "{project}" and all its task records?',
     "newTask.title": "What do you want to build today?",
     "newTask.instructionsMissing": "No {file} found in this project.",
-    "newTask.addInstructions": "Add a {file} to the project root to give {agent} context about your codebase, conventions, and preferences — it will follow them on every task.",
+    "newTask.addInstructions":
+      "Add a {file} to the project root to give {agent} context about your codebase, conventions, and preferences — it will follow them on every task.",
     "newTask.promptPlaceholder": "Describe your task… type @ to mention a file",
     "newTask.moreComposeActions": "More compose actions",
     "newTask.images": "Images",
     "newTask.planMode": "Plan mode",
     "newTask.send": "Send",
+    "newTask.newLine": "New line",
     "newTask.saveAsTodo": "Save as Todo",
     "newTask.imagesMustSend": "Tasks with images must be sent immediately.",
     "mention.loadingFiles": "Loading files…",
     "mention.allFiles": "all files",
     "mention.startTyping": "Start typing to search files",
-    "mention.noResults": "No results for \"{query}\"",
-    "mention.noFilesMatching": "No files matching \"{query}\"",
+    "mention.noResults": 'No results for "{query}"',
+    "mention.noFilesMatching": 'No files matching "{query}"',
     "mention.otherProjects": "Other Projects",
     "file.files": "Files",
     "file.emptyDirectory": "Empty directory",
@@ -277,7 +288,8 @@ const translations: Record<AppLanguage, Record<string, string>> = {
     "day.sat": "Sat",
     "toast.initProjectConfigFailed": "Failed to initialize project config: {error}",
     "toast.cancelTaskFailed": "Failed to cancel task: {error}",
-    "toast.loadProjectFilesFailed": "Failed to load project file list, @ references unavailable: {error}",
+    "toast.loadProjectFilesFailed":
+      "Failed to load project file list, @ references unavailable: {error}",
     "toast.saveProjectsFailed": "Failed to save project list: {error}",
     "toast.saveTasksFailed": "Failed to save tasks for project {projectId}: {error}",
   },
@@ -285,6 +297,7 @@ const translations: Record<AppLanguage, Record<string, string>> = {
     "appSettings.title": "应用设置",
     "appSettings.general": "通用",
     "appSettings.theme": "主题",
+    "appSettings.shortcuts": "快捷键",
     "appSettings.about": "关于",
     "appSettings.language": "语言",
     "appSettings.appLanguage": "应用语言",
@@ -306,6 +319,11 @@ const translations: Record<AppLanguage, Record<string, string>> = {
     "appSettings.starHint": "如果你觉得 NeZha 有帮助，欢迎在 GitHub 上给项目点星。",
     "appSettings.syntaxHighlightUnavailable": "语法高亮暂不可用，正在显示纯文本。",
     "appSettings.configFileNotFound": "未找到配置文件：{path}",
+    "appSettings.sendMessage": "发送消息",
+    "appSettings.sendShortcutModEnter": "Cmd/Ctrl + Enter",
+    "appSettings.sendShortcutModEnterHint": "使用 {shortcut} 发送，Enter 会换行。",
+    "appSettings.sendShortcutEnter": "Enter",
+    "appSettings.sendShortcutEnterHint": "使用 Enter 发送，Cmd/Ctrl + Enter 会换行。",
     "theme.followSystem": "跟随系统",
     "theme.followSystemAria": "跟随系统主题",
     "theme.manualTheme": "手动主题",
@@ -410,12 +428,14 @@ const translations: Record<AppLanguage, Record<string, string>> = {
     "task.deleteProjectPrompt": "删除项目“{project}”及其所有任务记录？",
     "newTask.title": "今天想构建什么？",
     "newTask.instructionsMissing": "此项目中未找到 {file}。",
-    "newTask.addInstructions": "在项目根目录添加 {file}，为 {agent} 提供代码库、约定和偏好上下文；它会在每个任务中遵循这些内容。",
+    "newTask.addInstructions":
+      "在项目根目录添加 {file}，为 {agent} 提供代码库、约定和偏好上下文；它会在每个任务中遵循这些内容。",
     "newTask.promptPlaceholder": "描述你的任务… 输入 @ 可引用文件",
     "newTask.moreComposeActions": "更多编辑操作",
     "newTask.images": "图片",
     "newTask.planMode": "计划模式",
     "newTask.send": "发送",
+    "newTask.newLine": "换行",
     "newTask.saveAsTodo": "保存为待办",
     "newTask.imagesMustSend": "包含图片的任务必须立即发送。",
     "mention.loadingFiles": "正在加载文件…",
